LPS-72067 Move logic to avoid a minor version increase
