+++ modules/sdk/gradle-plugins-defaults/src/main/java/com/liferay/gradle/plugins/defaults/tasks/WriteFindBugsProjectTask.java
@@ -26,8 +26,6 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
-import java.util.Set;
-
 import org.gradle.api.DefaultTask;
 import org.gradle.api.Project;
 import org.gradle.api.file.FileCollection;
@@ -93,16 +91,6 @@ public class WriteFindBugsProjectTask extends DefaultTask {
 		return _srcDirs;
 	}
 
-	public boolean isClasspathEmpty() {
-		if (_classpath == null) {
-			return true;
-		}
-
-		Set<File> files = _classpath.getFiles();
-
-		return !_containsClassOrJar(files.toArray(new File[files.size()]));
-	}
-
 	public void setAuxClasspath(FileCollection auxClasspath) {
 		_auxClasspath = auxClasspath;
 	}
@@ -162,30 +150,6 @@ public class WriteFindBugsProjectTask extends DefaultTask {
 
 			bufferedWriter.write("</Project>");
 		}
-	}
-
-	private boolean _containsClassOrJar(File[] files) {
-		for (File file : files) {
-			if (!file.exists()) {
-				continue;
-			}
-
-			if (file.isFile()) {
-				String fileName = file.getName();
-
-				if (fileName.endsWith(".class") || fileName.endsWith(".jar")) {
-					return true;
-				}
-
-				return false;
-			}
-
-			if (_containsClassOrJar(file.listFiles())) {
-				return true;
-			}
-		}
-
-		return false;
 	}
 
 	private void _writeFileElements(
